Move runtime Printer class and relevant typedefs into their own header so it isn't necessary to have a duplicate declaration of halide_float16_t in ``runtime_internal.h``.
