+++ src/runtime/android_opengl_context.cpp
@@ -1,5 +1,6 @@
 #include "runtime_internal.h"
 #include "HalideRuntime.h"
+#include "printer.h"
 
 extern "C" {
