+++ src/runtime/float16_t.cpp
@@ -1,4 +1,5 @@
 #include "runtime_internal.h"
+#include "HalideRuntime.h"
 
 extern "C" {
