+++ src/runtime/opencl.cpp
@@ -2,6 +2,7 @@
 #include "scoped_spin_lock.h"
 #include "device_interface.h"
 #include "HalideRuntimeOpenCL.h"
+#include "printer.h"
 
 #include "mini_cl.h"
