+++ src/runtime/osx_opengl_context.cpp
@@ -1,5 +1,6 @@
 #include "runtime_internal.h"
 #include "HalideRuntime.h"
+#include "printer.h"
 #include "scoped_mutex_lock.h"
 
 #define USE_AGL 0
