+++ src/runtime/posix_error_handler.cpp
@@ -1,5 +1,7 @@
+// FIXME: The first two includes should probably be swapped!
 #include "HalideRuntime.h"
 #include "runtime_internal.h"
+#include "printer.h"
 
 namespace Halide { namespace Runtime { namespace Internal {
