+++ src/runtime/renderscript.cpp
@@ -1,6 +1,7 @@
 #include "runtime_internal.h"
 #include "device_interface.h"
 #include "HalideRuntimeRenderscript.h"
+#include "printer.h"
 
 extern "C" void *dlopen(const char *, int);
 extern "C" void *dlsym(void *, const char *);
