+++ src/runtime/tracing.cpp
@@ -1,5 +1,6 @@
 #include "runtime_internal.h"
 #include "HalideRuntime.h"
+#include "printer.h"
 #include "scoped_spin_lock.h"
 
 extern "C" {
